Feat: implementar funciones en proyect_users con validaciones de permisos por usuario

+++ src/services/project_users.service.js
@@ -3,21 +3,28 @@ import User from '../models/user.model.js'
 import Project from '../models/project.model.js'
 import Role from '../models/role.model.js'
 
-export const createMember = async (userId, projectId, roleId) => {
+export const createMember = async (requesterId, userId, projectId, roleId) => {
+  const verifyRequester = await Project_users.findOne({
+    where: { user_id: requesterId, project_id: projectId },
+    include: Role
+  })
+  if (!verifyRequester) throw new Error('Requester does not have permissions')
   const user = await User.findByPk(userId)
   if (!user) throw new Error('User not found')
-
-  const project = await Project.findByPk(projectId)
-  if (!project) throw new Error('Project not found')
-
-  const role = await Role.findByPk(roleId)
-  if (!role) throw new Error('Role not found')
-
+  const userRole = await Role.findByPk(roleId)
+  if (!userRole || userRole.name === 'owner') throw new Error('Role not found')
   const existingMember = await Project_users.findOne({
     where: { user_id: userId, project_id: projectId }
   })
   if (existingMember) throw new Error('The user already belongs to this project')
-
+  if (!(verifyRequester.Role.name === 'admin' || verifyRequester.Role.name === 'owner')) {
+    throw new Error('Requester does not have permissions')
+  }
+  if (verifyRequester.Role.name === 'admin') {
+    if (userRole.name === 'admin') {
+      throw new Error('Requester does not have permissions')
+    }
+  }
   const member = await Project_users.create({
     user_id: userId,
     project_id: projectId,
@@ -26,32 +33,79 @@ export const createMember = async (userId, projectId, roleId) => {
   return member
 }
 
-export const updateMember = async (userId, projectId, newRoleId) => {
-  const member = await findMember(userId, projectId)
-
-  const role = await Role.findByPk(newRoleId)
-  if (!role) throw new Error('Role not found')
-
+export const updateMember = async (requesterId, userId, projectId, newRoleId) => {
+  const verifyRequester = await Project_users.findOne({
+    where: { user_id: requesterId, project_id: projectId },
+    include: [{
+      model: Role,
+      attributes: ['name']
+    }]
+  })
+  if (!verifyRequester) throw new Error('Requester does not have permissions')
+  const user = await User.findByPk(userId)
+  if (!user) throw new Error('User not found')
+  const userRole = await Role.findByPk(newRoleId)
+  if (!userRole || userRole.name === 'owner') throw new Error('Role not found')
+  const member = await Project_users.findOne({
+    where: { user_id: userId, project_id: projectId },
+    include: [{
+      model: Role,
+      attributes: ['name']
+    }]
+  })
+  if (!member) throw new Error('Member not found')
+  if (!(verifyRequester.Role.name === 'admin' || verifyRequester.Role.name === 'owner')) {
+    throw new Error('Requester does not have permissions')
+  }
+  if (verifyRequester.Role.name === 'admin') {
+    if (member.Role.name === 'admin') {
+      throw new Error('Requester does not have permissions')
+    }
+  }
+  if (member.Role.name === 'admin') {
+    if (!verifyRequester.Role.name === 'owner') {
+      throw new Error('Requester does not have permissions')
+    }
+  }
   await member.update({
     role_id: newRoleId
   })
 }
 
-export const deleteMember = async (userId, projectId) => {
-  const member = await findMember(userId, projectId)
-  await member.destroy({ force: true })
-}
+export const deleteMember = async (requesterId, userId, projectId) => {
 
-const findMember = async (userId, projectId) => {
+  const verifyRequester = await Project_users.findOne({
+    where: { user_id: requesterId, project_id: projectId },
+    include: [{
+      model: Role,
+      attributes: ['name']
+    }]
+  })
+  if (!verifyRequester) throw new Error('Requester does not have permissions')
   const user = await User.findByPk(userId)
   if (!user) throw new Error('User not found')
-
-  const project = await Project.findByPk(projectId)
-  if (!project) throw new Error('Project not found')
-
+  // const userRole = await Role.findByPk(newRoleId)
+  // if (!userRole || userRole.name === 'owner') throw new Error('Role not found')
   const member = await Project_users.findOne({
-    where: { user_id: userId, project_id: projectId }
+    where: { user_id: userId, project_id: projectId },
+    include: [{
+      model: Role,
+      attributes: ['name']
+    }]
   })
-  if (!member) throw new Error('Membership not found')
-  return member
+  if (!member) throw new Error('Member not found')
+  if (!(verifyRequester.Role.name === 'admin' || verifyRequester.Role.name === 'owner')) {
+    throw new Error('Requester does not have permissions')
+  }
+  if (verifyRequester.Role.name === 'admin') {
+    if (member.Role.name === 'admin' || member.Role.name === 'owner') {
+      throw new Error('Requester does not have permissions')
+    }
+  }
+  if (member.Role.name === 'admin') {
+    if (!verifyRequester.Role.name === 'owner') {
+      throw new Error('Requester does not have permissions')
+    }
+  }
+  await member.destroy({ force: true })
 }
